% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAUtilities.r
\name{AA_Rand}
\alias{AA_Rand}
\title{Random return(s) from a lognormally distribution}
\usage{
AA_Rand(n, r, sd, t, seed = NA)
}
\arguments{
\item{n}{Number of returns to generate}

\item{r}{Mean return}

\item{sd}{Standard deviation of returns}

\item{t}{Time in years}

\item{seed}{Random seed}
}
\value{
value Random returns
}
\description{
Returns n lognormally distributed random variables with a mean of r and a standard deviation of
sd.  Periods over 1 year are annualized.  Specify inputs in percentage formats (e.g. 8%=8, not 0.08).
}
\examples{
AA_Rand(10,.08,12,10)

}

