% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAUtilities.r
\name{AA_Return}
\alias{AA_Return}
\title{Lognormally distributed return}
\usage{
AA_Return(p, r, sd, t)
}
\arguments{
\item{p}{Probability of doing worse than result}

\item{r}{Mean return}

\item{sd}{Standard deviation of returns}

\item{t}{Period in years}
}
\value{
value Return
}
\description{
Calculates the return for which there is only a p chance of doing worse
than over t years given a mean of r and a standard deviation of
sd. Periods over 1 year are annualized. Specify inputs in percentage formats (e.g. 8%=8, not 0.08).
}
\examples{
AA_Return(.5,8,12,10)

}

