% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAUtilities.r
\name{diversification_score}
\alias{diversification_score}
\title{Diversification score}
\usage{
diversification_score(w)
}
\arguments{
\item{w}{Weights of a portfolio}
}
\value{
value Diversification score
}
\description{
The diversification score for a portfolio of n equally-weighted assets will
be n.  All things being equal, the larger the value the greater the
diversification. This is a simple measure ignoring the covariance matrix. Use
with caution.
}
\examples{
diversification_score(c(.5,.4,.1))
}

